Project refactoring : MVVM folder structure
In order to simply the development following the MVVM design pattern, I changed the structure of the directories in the project.
Please explore it carefully and put your new screens in the screen folder
