+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:label="@string/title_activity_second"
             android:theme="@style/Theme.Lasta" />
         <activity
-            android:name="com.lastaoutdoor.lasta.MainActivity"
+            android:name="com.lastaoutdoor.lasta.view.MainActivity"
             android:exported="true"
             android:label="@string/title_activity_main"
             android:theme="@style/Theme.Lasta">
