--- app/src/main/java/com/lastaoutdoor/lasta/resources/C.kt
@@ -1,12 +1,0 @@
-package com.lastaoutdoor.lasta.resources
-
-// Like R, but C
-object C {
-  object Tag {
-    const val greeting = "main_screen_greeting"
-    const val greeting_robo = "second_screen_greeting"
-
-    const val main_screen_container = "main_screen_container"
-    const val second_screen_container = "second_screen_container"
-  }
-}
